rename and move authentication error hints
